Add support for Nova translatable and user locale
In this commit, Nova translatable package was added to the dependencies in composer.json. Also, a locale field was added to the 'users' table for storing users' language preferences. User locale is now updated based on authentication status and language setting in the NovaLocalizationServiceProvider file. Furthermore, right-to-left support for Nova was enabled for the Arabic language.